extract overlapping constants into a new monorepo package that both the api and web packages can depend on
