+++ packages/constants/src/error-codes.js
@@ -1,0 +1,5 @@
+export const errorCodes = Object.freeze({
+  PICKING: "PICKING",
+  NO_TITLE: "NO_TITLE",
+  NO_LIST_NAME: "NO_LIST_NAME",
+});
