+++ packages/constants/src/ratings.js
@@ -1,0 +1,11 @@
+export const ratingsSource = Object.freeze({
+  IMDB: "IMDB",
+  ROTTEN_TOMATOES: "ROTTEN_TOMATOES",
+  METACRITIC: "METACRITIC",
+});
+
+export const ratingsSources = [
+  ratingsSource.IMDB,
+  ratingsSource.ROTTEN_TOMATOES,
+  ratingsSource.METACRITIC,
+];
