+++ packages/web/index.html
@@ -4,6 +4,7 @@
     <title>Movie Decider 4000</title>
     <link rel="icon" href="/favicon.ico" />
     <link rel="apple-touch-icon" href="touch-icon-iphone.png" />
+    <meta name="viewport" content="width=device-width" />
     <meta name="apple-mobile-web-app-title" content="MD4K" />
     <meta name="apple-mobile-web-app-capable" content="yes" />
     <meta name="apple-mobile-web-app-status-bar-style" content="black" />
