+++ packages/web/package.json
@@ -25,6 +25,7 @@
     "date-fns": "^2.29.2",
     "graphql": "^16.5.0",
     "lodash": "^4.17.21",
+    "md4k-constants": "*",
     "omit-deep-lodash": "^1.1.7",
     "react": "^18.2.0",
     "react-day-picker": "8.0.0-beta.29",
