+++ packages/web/src/components/app/components/list/components/ratings/ratings.jsx
@@ -1,9 +1,7 @@
 import React from "react";
 
-import {
-  ratingsSource,
-  ratingsSourceImage,
-} from "../../../../../../constants/ratings";
+import { ratingsSource } from "md4k-constants";
+import { ratingsSourceImage } from "../../../../../../constants/ratings";
 import {
   denseMargins,
   RatingsList,
