--- packages/web/src/constants/genres.js
@@ -1,27 +1,0 @@
-export const genres = {
-  NONE: 0,
-  COMEDY: 1,
-  DRAMA: 2,
-  ACTION: 3,
-  SCI_FI: 4,
-  ADVENTURE: 5,
-  FAMILY: 6,
-  THRILLER: 7,
-  MUSICAL: 8,
-  DOCUMENTARY: 9,
-  HORROR: 10,
-};
-
-export const genreLabels = {
-  [genres.NONE]: "None",
-  [genres.COMEDY]: "Comedy",
-  [genres.DRAMA]: "Drama",
-  [genres.ACTION]: "Action",
-  [genres.SCI_FI]: "Sci-Fi",
-  [genres.ADVENTURE]: "Adventure",
-  [genres.FAMILY]: "Family",
-  [genres.THRILLER]: "Thriller",
-  [genres.MUSICAL]: "Musical",
-  [genres.DOCUMENTARY]: "Documentary",
-  [genres.HORROR]: "Horror",
-};
